+++ student-blogs/cesarurbina-4941/package.json
@@ -1,0 +1,23 @@
+{
+  "name": "exercise-2",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
+    "sass": "sass src/main.scss src/styles/main.css"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "dayjs": "^1.11.8"
+  },
+  "devDependencies": {
+    "sass": "^1.63.6",
+    "vite": "^4.3.9"
+  }
+}
