--- student-cvs/cesarurbina-4941/exercise-1/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "exercise-1",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "type": "module",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "lodash": "^4.17.21"
-  }
-}
