--- student-cvs/cesarurbina-4941/exercise-2/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "exercise-2",
-  "version": "1.0.0",
-  "description": "",
-  "main": "main.mjs",
-  "type": "module",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "dayjs": "^1.11.8"
-  }
-}
