--- student-cvs/cesarurbina-4941/exercise-3/main.mjs
@@ -1,8 +1,0 @@
-// main.mjs
-function handleSubmit(event) {
-  event.preventDefault(); // 🍎
-  console.log("Hola mundo");
-}
-
-const $form = document.querySelector("form");
-$form.addEventListener("submit", handleSubmit);
