--- student-cvs/cesarurbina-4941/exercise-3/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "exercise-2",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "vite",
-    "build": "vite build",
-    "preview": "vite preview",
-    "sass": "sass src/main.scss src/styles/main.css"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "sass": "^1.63.6",
-    "vite": "^4.3.9"
-  }
-}
