--- student-cvs/cesarurbina-4941/exercise-3/src/base.scss
@@ -1,234 +1,0 @@
-@import "./variables.scss";
-
-html, body {
-    height: 100%;
-    box-sizing: border-box;
-}
-
-* {
-    box-sizing: inherit;
-}
-
-body {
-    background-color: $athensGray;
-    margin: 0%;
-    font-family: 'Didact Gothic', sans-serif;
-}
-
-.logo {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    padding: 0.8rem;
-}
-
-.logo span {
-    font-family: 'Didact Gothic', sans-serif;
-    font-size: 20px;
-    color: $fiord;
-}
-
-.nav {
-    background-color: $athensGrayLite;
-    height: 2rem;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-
-    margin-left: 1.3rem;
-    border-left: 5px solid $flamingo;
-}
-
-.nav-list {
-    display: flex;
-    list-style-type: none;
-    padding: 0;
-    margin: 0;
-}
-
-.item a {
-    margin: 0 0.6rem;
-    color: $ebony;
-    text-decoration: none;
-    text-transform: uppercase;
-    font-size: 14px;
-    letter-spacing: 1px;
-    font-family: 'Didact Gothic', sans-serif;
-}
-
-.container {
-    display: flex;
-    flex-direction: column;
-}
-
-.container .header {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.container .header h1 {
-    margin: 0.6rem;
-}
-
-.container .span-hello {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    font-size: 100px;
-    font-weight: normal;
-    height: 100px;
-    font-family: Helvetica, sans-serif;
-    color: $flamingo;
-}
-
-.container .span-name {
-    display: flex;
-    justify-content: flex-start;
-    align-items: center;
-    padding-left: 0.3rem;
-    height: 50px;
-    font-size: 45px;
-    font-weight: normal;
-    font-family: 'Didact Gothic', sans-serif;
-    color: $ebony;
-}
-
-.container .study {
-    display: flex;
-    justify-content: flex-start;
-    padding-left: 0.5rem;
-    margin: 0 0 0 1.1rem;
-    font-size: 16px;
-    letter-spacing: 1px;
-    font-family: 'Didact Gothic', sans-serif;
-    color: $ebony;
-    border-left: 5px solid $flamingo;
-}
-
-.container .header-about {
-    margin: 0.5rem 1rem 0 1rem;
-}
-
-.container .header-about p {
-    font-size: 16px;
-    text-align: justify;
-    word-spacing: -2px;
-    color: $ebony;
-    letter-spacing: 1px;
-}
-
-.container .header-about h2, h3 {
-    color: $ebony;
-    letter-spacing: 1px;
-    font-family: 'Didact Gothic', sans-serif;
-    font-weight: normal;
-}
-
-.container .header-about h2 {
-    margin-top: 1rem;
-    margin-bottom: 0.5rem;
-    font-size: 20px;
-}
-
-.container .header-about h3 {
-    margin-top: 0;
-    font-size: 16px;
-}
-
-.container .commands {
-    background-color: $athensGrayLite;
-    padding: 0;
-    width: 94.5%;
-
-    border-right: 4px solid $corn;
-}
-
-.container li {
-    margin: 0.6rem 0;
-}
-
-.container li::marker {
-    color: $corn;
-}
-
-.container code {
-    background-color: $athensGray;
-    padding: 0.3em 0.5em;
-}
-
-.container .personal-goals {
-    margin: 1.2rem 1rem 0 1rem;
-}
-
-.container .personal-goals h2 {
-    color: $ebony;
-    letter-spacing: 1px;
-    font-family: 'Didact Gothic', sans-serif;
-    font-weight: normal;
-
-    margin-top: 0.2rem;
-    margin-bottom: 0.5rem;
-    font-size: 20px;
-}
-
-.container .personal-goals p {
-    margin-top: 0;
-    font-size: 16px;
-    text-align: justify;
-    word-spacing: -2px;
-    color: $ebony;
-    letter-spacing: 1px;
-}
-
-.container .canada {
-    margin: 0.6rem 1rem 0 1rem;
-}
-
-.container .canada img {
-    width: 100%;
-    height: auto;
-}
-
-.container .block-while-green {
-    height: 2rem;
-    width: 100%;
-    margin-left: 49.5%;
-    margin-bottom: 0.5rem;
-    background-color: $athensGrayLite;
-
-    border-left: 4px solid $lima;
-}
-
-.container .block-while-blue {
-    height: 2rem;
-    width: 50.5%;
-    margin-right: 2rem;
-    margin-top: 1.1rem;
-    margin-bottom: 1.1rem;
-    background-color: $athensGrayLite;
-
-    border-right: 4px solid $lochmara;
-}
-
-.container .norway {
-    margin: 0 1rem 0.4rem 1rem;
-}
-
-.container .norway img {
-    width: 100%;
-    height: auto;
-}
-
-.container .footer {
-    display: flex;
-    justify-content: flex-start;
-    align-items: center;
-    height: 2rem;
-
-    background-color: $athensGrayLite;
-}
-
-.container .footer span {
-    margin-left: 1rem;
-}
