--- student-cvs/cesarurbina-4941/exercise-3/src/desktop.scss
@@ -1,312 +1,0 @@
-@import "./variables.scss";
-
-/* Responsive - Mobile first*/
-
-@media screen and (min-width: 251px) {
-
-    .container .canada {
-        display: flex;
-        justify-content: center;
-        align-content: center;
-    }
-
-    .container .canada img {
-        max-width: 600px;
-    }
-
-    .container .norway {
-        display: flex;
-        justify-content: center;
-        align-content: center;
-    }
-
-    .container .norway img {
-        max-width: 600px;
-    }
-}
-
-@media screen and (min-width: 371px) {
-    .container .commands {
-        width: 50.5%;
-    }
-}
-
-@media screen and (min-width: 618px) {
-    .container {
-        display: grid;
-        grid-template-columns: repeat(2, 1fr);
-        grid-template-rows: 13.4rem auto auto auto auto auto;
-        gap: 1rem 0;
-    }
-
-    .container .header {
-        align-items: normal;
-        padding-top: 0.3rem;
-    }
-
-    .container .header-about {
-        grid-column: 2/3;
-        grid-row: 1/2;
-
-        margin: 0;
-        padding: 0.3rem 1rem;
-    }
-
-    .container .commands {
-        grid-column: 2/3;
-        grid-row: 2/3;
-
-        width: 100%;
-        margin-left: 1rem;
-        border-right: none;
-        border-left: 4px solid $corn;
-        border-bottom: 4px solid $corn;
-    }
-
-    .container .block-while-green {
-        grid-column: 1/2;
-        grid-row: 2/3;
-
-        width: 95%;
-        margin: 0;
-        height: 2.8rem;
-        border-left: none;
-        border-right: 4px solid $lima;
-        border-top: 4px solid $lima;
-        background-color: transparent;
-    }
-
-    .container .block-while-blue {
-        width: 95%;
-        height: 6rem;
-    }
-
-    .container .blue-container {
-        grid-column: 2/3;
-        grid-row: 3/4;
-
-        display: flex;
-        justify-content: center;
-        align-items: center;
-    }
-
-    .container .personal-goals {
-        grid-column: 1/2;
-        grid-row: 2/3;
-
-        padding-top: 0.4rem;
-        margin: 2.5rem 2rem 0 2rem;
-    }
-
-    .container .canada {
-        margin: 0;
-    }
-
-    .container .norway {
-        grid-column: 1/3;
-        grid-row: 4/6;
-        margin:0 1rem;
-    }
-
-    .container .norway img {
-        max-width: 800px;
-    }
-
-    .container .footer {
-        grid-column: 1/3;
-        grid-row: 6/7;
-    }
-}
-
-@media screen and (min-width: 671px) {
-    .container .header-about {
-        padding: 0.9rem 1rem 0 1rem;
-    }
-
-    .container .commands {
-        margin-left: 1rem;
-    }
-}
-
-@media screen and (min-width: 852px) {
-    .container .header-about {
-        padding: 1rem 2rem 0 1rem;
-    }
-
-    .container .personal-goals p{
-        margin: 1rem 1rem 0 0;
-    }
-
-    .container .commands {
-        margin-left: 1rem;
-    }
-}
-
-@media screen and (min-width: 883px) {
-    .container .header-about {
-        padding: 0.7rem 2rem 0 1rem;
-    }
-
-    .container .header-about p {
-        font-size: 17px;
-    }
-
-    .container .personal-goals {
-        margin: 2rem 2rem 0 2rem;
-    }
-
-    .container .personal-goals p {
-        font-size: 17px;
-    }
-}
-
-@media screen and (min-width: 883px) {
-    .container .header-about {
-        padding: 1.4rem 2rem 0 1rem;
-    }
-}
-
-@media screen and (min-width: 1118px) {
-    .nav-container {
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        background-color: athensGrayLite;
-        height: 2.2rem;
-    }
-
-    .logo {
-        padding: 0 0 0 1rem;
-    }
-    
-    .logo span {
-        text-transform: uppercase;
-        letter-spacing: 1px;
-        font-size: 14px;
-    }
-
-    .nav-container {
-        background-color: $athensGrayLite;
-        height: 2rem;
-    }
-
-    .nav {
-        background-color: transparent;
-        border-left: none;
-        margin-right: 1rem;
-        margin-left: 0;
-        height: 2.2rem;
-    }
-    
-    .item a {
-        color: fiord;
-    }
-
-    .container {
-        display: grid;
-        grid-template-columns: repeat(2, 1fr);
-        grid-template-rows: 11.3rem 11.3rem 10rem auto;
-        gap: 0 1rem;
-    }
-
-    .container .header {
-        grid-column: 1/2;
-        grid-row: 1/3;
-    }
-
-    .container .header h1 {
-        margin-top: 1.3rem;
-    }
-
-    .container .span-hello {
-        font-size: 170px;
-        height: 170px;
-    }
-    
-    .container .span-name {
-        height: 90px;
-        font-size: 65px;
-        padding-left: 0.7rem;
-    }
-
-    .container .study {
-        margin: 0 0 0 1.6rem;
-        font-size: 15px;
-    }
-
-    .container .header-about {
-        padding: 0.8rem 2rem 0 0;
-    }
-
-    .container .header-about p {
-        font-size: 17px;
-    }
-
-    .container .header-about h2 {
-        margin: 0 0 0.2rem 0;
-    }
-
-    .container .commands {
-        grid-column: 2/3;
-        grid-row: 2/3;
-        margin-left: 0;
-    }
-
-    .container .block-while-green {
-        grid-column: 1/2;
-        grid-row: 3/4;
-
-        width: 100%;
-    }
-
-    .container .personal-goals {
-        grid-column: 1/2;
-        grid-row: 3/4;
-
-        margin: auto 2rem;
-        
-    }
-
-    .container .personal-goals p {
-        font-size: 18px;
-    }
-
-    .container .canada {
-        grid-column: 2/3;
-        grid-row: 4/5;
-
-        height: 50%;
-        width: 50%;
-    }
-
-    .container .norway {
-        grid-column: 1/2;
-        grid-row: 4/5;
-
-        width: auto;
-        height: auto;
-        margin: 0;
-    }
-}
-
-@media screen and (min-width: 1218px) {
-    .container .span-hello {
-        font-size: 200px;
-        height: 170px;
-    }
-    
-    .container .span-name {
-        height: 90px;
-        font-size: 80px;
-        padding-left: 0.7rem;
-    }
-
-    .container .study {
-        margin: 0 0 0 1.6rem;
-        font-size: 18px;
-    }
-
-    .container .header-about p {
-        font-size: 18px;
-    }
-}
