--- student-cvs/cesarurbina-4941/exercise-3/src/main.scss
@@ -1,3 +1,0 @@
-@use "../src/base.scss";
-@use "../src/desktop.scss";
-
