--- student-cvs/cesarurbina-4941/exercise-3/src/variables.scss
@@ -1,8 +1,0 @@
-$athensGray: #e5e7eb;
-$athensGrayLite: #f3f4f6;
-$fiord: #475569;
-$flamingo: #ef4444;
-$ebony: #0f172a;
-$corn: #eab308;
-$lima: #84cc16;
-$lochmara: #0284c7;
