+++ student-cvs/cesarurbina-4941/student-cvs/index.html
@@ -1,0 +1,96 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta
+      name="viewport"
+      content="width=device-width, initial-scale=1.0, maximum-scale=1.0"
+    />
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+    <link
+      href="https://fonts.googleapis.com/css2?family=Didact+Gothic&display=swap"
+      rel="stylesheet"
+    />
+    <link rel="stylesheet" href="/src/styles/main.css" />
+    <title>cesarurbina-4941</title>
+  </head>
+  <body>
+    <div class="nav-container">
+      <div class="logo">
+        <header>
+          <span>cesarurbina-4941 - responsive</span>
+        </header>
+      </div>
+      <div class="nav">
+        <nav>
+          <ul class="nav-list">
+            <li class="item"><a href="/index.html">Home</a></li>
+            <li class="item"><a href="/contact.html">Contact</a></li>
+          </ul>
+        </nav>
+      </div>
+    </div>
+    <div class="container">
+      <div class="header">
+        <header>
+          <h1>
+            <span class="span-hello">Hello!</span>
+            <span class="span-name">I am César</span>
+          </h1>
+          <p class="study">Student of Systems Engineering</p>
+        </header>
+      </div>
+      <div class="header-about">
+        <header>
+          <p>
+            I am from Colombia and I am 24 years old, I am currently completing
+            my university studies. I hope to improve my software development
+            skills this year.
+          </p>
+        </header>
+        <section>
+          <h2>My Terminal</h2>
+          <article class="commands-common">
+            <h3>My most used terminal commands.</h3>
+          </article>
+        </section>
+      </div>
+      <div class="commands">
+        <ul>
+          <li><code>cd</code></li>
+          <li><code>pwd</code></li>
+          <li><code>clear</code></li>
+          <li><code>ls</code></li>
+          <li><code>code .</code></li>
+        </ul>
+      </div>
+
+      <div class="personal-goals">
+        <h2 class="goals">Personal Goals</h2>
+        <p>
+          My most ambitious personal goal is to be able to migrate and live
+          several years in Canada and Norway
+        </p>
+      </div>
+      <div class="block-while-green"></div>
+      <div class="canada">
+        <img src="/static/canada.jpg" alt="canada" />
+      </div>
+      <div class="blue-container">
+        <div class="block-while-blue"></div>
+      </div>
+
+      <div class="norway">
+        <img src="/static/norway.jpg" alt="norway" />
+      </div>
+
+      <div class="footer">
+        <footer>
+          <span>Copyright 2023</span>
+        </footer>
+      </div>
+    </div>
+  </body>
+</html>
